+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.7
+
+* added Permissions Table in Method Dart Doc
+* updated Readme
+* added Contacts and ContactFolders Endpoints and Models
+
 ## 0.0.6
 
 * added Endpoints to Create Notebooks